refactor: 공통 도메인을 별도 패키지로 분리하라
