+++ src/main/java/kitchenpos/core/domain/Name.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.domain;
+package kitchenpos.core.domain;
 
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
