+++ src/main/java/kitchenpos/core/domain/Price.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.domain;
+package kitchenpos.core.domain;
 
 import java.math.BigDecimal;
 import java.util.Objects;
