+++ src/main/java/kitchenpos/menu/domain/Menu.java
@@ -9,6 +9,8 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.Table;
+import kitchenpos.core.domain.Name;
+import kitchenpos.core.domain.Price;
 
 @Entity
 @Table(name = "menu")
